Tweak OpenAPI Definitions (#39)
You can externally reference other schemas in OpenAPI, so we can share
things across different microservices (yay), to do this you need to
import the generated types and schema, so this puts that in place.
Secondly we need to tweak the read metadata to add additional scope
where it's required for various views.  Third, we can share some
metadata conversion functions.