Add Shared CA to Chart (#53)
One of the bothersome issues with deploying a development cluster is the
issue of CAs.  At present all components have a self-signed issuer,
which means you have to import a bunch of CAs into your browser.  This
allows the generation of a centralized CA (or the import of one from
mkcert for example) so you only have to do this once.  This also changes
the certificate propagation to services by forcing the referencing of a
TLS secret.