Expand Scoped Metadata
In views of all clusters in an organization for example, it would be
nice to see scoping information such as the project name in order to
differentiate resources.  At present it's possible doing a project
lookup, but rather than use another API call, we could just return this
in the metadata.  Obviously this view is point in time, so you'll need
to poll to see updates.