Fix Userinfo
This is purely a token introspection type now, we are no longer piggy
backing RBAC on it, instead prefering the ACL interface.  A vast part of
this is moving RBAC and authorization into identity to save pain and
suffering.  A little bit of coupling still exists in order to get access
to the request subject for auditing.