Propagate Remote With Context (#72)
We already do this with Kubernetes clients, where they are attached to
the context and propagated to descendant provisioners so the scope
magically updates as we traverse the call graph.  Do the same for the
remote, which a) makes things cleaner and we don't need to manually
propagate to child metadata, and b) gives you access to a whole host
more information e.g. the Kubernetes endpoint for troublesome
provisioners such as Cilium.  While we are here, remove the similiar
abomination that this background deletion.