Add a Client Trust Anchor (#74)
We're using RFC8705 to issue oauth2 access tokens to services (e.g.
kubernetes) so they can talk to other APIs (e.g. region).  This requires
the use of mTLS, and for that we need a trust anchor to pass to the TLS
termination. On top of all that we need to rationalize client TLS
handling, remove all the identity specific bits from core - it's just a
pain in arse, allow reconcilers to register and use CLI flags.