Refactor CA Logic (#84)
The way this was done was a little half-baked, everything was separate
except the self-signed issuer, which would lead to really confusing
logic to enable it based on both whether it was required by the client
CA or server CA, compounded by the fact the server CA needed to generate
it in different circumstanes whether or not the CA was provided from an
external source.  This separates concerns, so the logic is vastly
simplified, and makes client CA stuff un-conditional, as they are all
internal anyway and not accessed by a browser.