Level Up Semantic Versions (#87)
We should treat these better and use a proper parsed backing type.  We
then need to add constraints to applications so the resolver can work in
a nicer fashion.  This also takes the liberty of upgrading libs so I can
use k8s native functions, but that requires a new compiler and linter,
which is good as it fixes CVEs.