Add New Resource Labels
Quota allocations for example are owned by a resource e.g. a Kubernetes
cluster.  We need to be able to see what allocation is owned by what
resource, and also have a performant way of looking up the allocation
that's onwed by a resource for updates etc. Thus labels!  Usually we'd
use a Kubernetes core object reference, but to be flexible we envisage
quotas being defined by 3rd party services too, so keep it flexible.