Give provisioners a way to supply namespace labels
 - give the driver-neutral cd.HelmApplication a field for (created)
   namespace metadata;
 - translate that into the argov1.Application equivalent;
 - when generating a cd.HelmApplication, look for an optional
   NamespaceLabeler interface on the generator, in the same way as for
   values

This is mostly useful for virtual kubernetes clusters (in
uni-kubernetes), since those get applied in their own namespace on the
remote cluster, and it's useful to be able to link them back to their
origin.

In the absence of the interface, everything should be generated just
the same, so this will be backward-compatible.