Give provisioners a way to supply namespace labels (#149)
* Give provisioners a way to supply namespace labels

 - give the driver-neutral cd.HelmApplication a field for (created)
   namespace metadata;
 - translate that into the argov1.Application equivalent;
 - when generating a cd.HelmApplication, look for an optional
   NamespaceLabeler interface on the generator, in the same way as for
   values

This is mostly useful for virtual kubernetes clusters (in
uni-kubernetes), since those get applied in their own namespace on the
remote cluster, and it's useful to be able to link them back to their
origin.

In the absence of the interface, everything should be generated just
the same, so this will be backward-compatible.

* Test that namespace metadata are passed to the driver

This commit incorporates the expected call of
NamespaceLabeler.NamespaceMetadata(...) into the provisioner
tests. It'll fail if that method isn't called or its return values
aren't used properly, since the argument the driver mock expects now
includes labels and annotations from the generator.

* Alter cd/driver/argocd test to cover namespace metadata

This demonstrates:

 - if you don't supply HelmApplication.NamespaceMetadata, then
   Application.SyncPolicy.ManagedNamespaceMetadata stays nil (i.e.,
   the prior behaviour); - if you do supply it, then the bit that gets
   supplied is present in the Application field, and the bit that
   doesn't, isn't. (i.e., the new behaviour)

I've also made a type for HelmApplication.NamespaceMetadata, because
it's otherwise painful using literal syntax in tests.