Service Messaging
Services may need to provision infrastructure in their own namespace on
behalf of a user.  This is primarily a security boundary so a user does
not have access to the underlying infrastructure (e.g. via a kubeconfig
for a Kubernetes based service), due to elevated privileges for example.

However we want the user (principal) to be available in the Kubernetes
service so that it correctly attributes the resource allocations to the
right organization and project.  For that we need a secure mechanism to
relay principal identity information to other services.  Luckily, as all
services talk to one another with mTLS, we already have a secure key
distribution mechanism that end users cannot use (they don't have access
to the CA's private key, so cannot invent identities that can be
trusted).  Thus we provide a couple interfaces that allow arbitrary data
to be encoded and signed into a suitable form for injection as a HTTP
header, and allow a HTTP header to be verified and decoded again on the
other end.